zjut fix warnings 22.12.18

+++ mindspore/core/ops/lgamma.cc
@@ -57,7 +57,7 @@ TypePtr LgammaInferType(const PrimitivePtr &primitive, const std::vector<Abstrac
 AbstractBasePtr LgammaInfer(const abstract::AnalysisEnginePtr &, const PrimitivePtr &primitive,
                             const std::vector<AbstractBasePtr> &input_args) {
   MS_EXCEPTION_IF_NULL(primitive);
-  const size_t input_num = 1;
+  const int64_t input_num = 1;
   CheckAndConvertUtils::CheckInputArgs(input_args, kEqual, input_num, primitive->name());
   auto shapes = LgammaInferShape(primitive, input_args);
   auto types = LgammaInferType(primitive, input_args);
